What is the point of doing frontents, if you cant make them look nice
The classes with the animations are clearly assigned to the nodes, so why are the animations
not playing?

I give up. Angular wants me to do animations their way, but this thing, that takaes just 3 or 4
lines of css requires a rediculess setup in angular all because this stupid angular renderer
only updates the view, when it feels like doing so and not when actual properties change.

Appearently I cant have both, reactivity and animations, all because the reactive part requires
the dom to be redrawn, but it uses the same components. So it does not even make sense, why the
CSS-Animations are not playing.